backend/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
langchain==1.0.2
langchain-community==0.4
langchain-huggingface==1.0.0
sentence-transformers==5.1.2
pydantic==2.12.3
python-dotenv==1.1.1

# cloud
google-cloud-compute==1.40.0
google-cloud-storage==3.4.1
google-cloud-monitoring==2.28.0
boto3==1.40.58
azure-identity==1.25.1
azure-mgmt-compute==37.0.1
azure-mgmt-storage==24.0.0
azure-mgmt-monitor==7.0.0
azure-storage-blob==12.27.0
ibm-vpc==0.31.0
ibm-cos-sdk==2.14.3

# llm
langchain-google-genai==3.0.0
langchain-openai==1.0.1
# ibm-watsonx-ai==1.4.2
# langchain-ibm==0.3.19